refactor(backend): split proxy and upstream account routing modules
