+++ docs/specs/README.md
@@ -163,3 +163,4 @@
 | huzqt | 前端结构收敛 follow-up | 已完成（PR #331） | `huzqt-frontend-structure-convergence-followup/SPEC.md` | 2026-04-11 | fast-track / account-pool page decomposition + typed follow-up / storybook evidence shown / shared-testbox browser smoke green / merged + cleanup |
 | phb37 | 后端结构收敛 follow-up | 已完成 | `phb37-backend-structure-convergence-followup/SPEC.md` | 2026-04-11 | fast-track / maintenance archive + hourly-rollup helpers converged into real support modules / local cargo fmt+check+test green / shared-testbox proxy+raw smoke green |
 | ts6qp | account-pool 前端类型债清偿 | 已完成 | `ts6qp-account-pool-typing-debt/SPEC.md` | 2026-04-12 | normal-flow / production `@ts-nocheck` removed from account-pool pages / `UpstreamAccounts` helper extraction + Create strict typing landed / local lint+test+build+storybook green / visual evidence archived |
+| 9vau7 | 后端结构债双 PR 快车道 | 待实现 | `9vau7-backend-structure-dual-pr-followup/SPEC.md` | 2026-04-12 | fast-track / two sequential backend PRs / PR1=`proxy+upstream_accounts/routing` / PR2=`api+router builders+api-read smoke` / merge+cleanup |
