+++ src/upstream_accounts/mod.rs
@@ -1,12 +1,19 @@
 //! Upstream accounts domain entrypoint split into focused slices.
-//!
-//! Keep include order stable to preserve item resolution and behavior parity.
+
+use crate::*;
 
 include!("core.rs");
-include!("crud_group_notes.rs");
-include!("imports_jobs_sse.rs");
-include!("oauth_sessions_callbacks.rs");
-include!("maintenance_dispatch.rs");
+mod crud_group_notes;
+mod imports_jobs_sse;
+mod maintenance_dispatch;
+mod oauth_sessions_callbacks;
+mod routing;
 include!("sync.rs");
-include!("routing.rs");
-include!("tests.rs");
+#[cfg(test)]
+mod tests;
+
+pub(crate) use crud_group_notes::*;
+pub(crate) use imports_jobs_sse::*;
+pub(crate) use maintenance_dispatch::*;
+pub(crate) use oauth_sessions_callbacks::*;
+pub(crate) use routing::*;
