+++ src/upstream_accounts/routing/mod.rs
@@ -1,0 +1,13 @@
+use super::*;
+
+mod internal {
+    use super::*;
+
+    include!("settings_runtime.rs");
+    include!("candidate_loading.rs");
+    include!("selection.rs");
+    include!("failure_recording.rs");
+    include!("sticky_routes.rs");
+}
+
+pub(crate) use internal::*;
