+++ src/upstream_accounts/tests_part_1.rs
@@ -1,3 +1,4 @@
+    #[allow(unused_imports)]
     use super::*;
     use axum::{
         Json, Router,
